Changed return type of division function to 32-bit integer

+++ FastBit Courses/Cortex M3 and M4 processors/My_Workspace/MathOperation_usingSVC/Src/math.c
@@ -16,8 +16,7 @@ extern int32_t subtraction(int32_t a, int32_t b){
 extern int32_t multiplication(int32_t a, int32_t b){
 	return a * b;
 }
-extern float division(int32_t a, int32_t b){
-	/* Enable floating point unit */
+extern int32_t division(int32_t a, int32_t b){
 	return a / b;
 }
